Remove Sets
These are just too generic and deserve to live in their own repo.
