+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/spjmurray/go-sat
 
 go 1.23.2
+
+require github.com/spjmurray/go-util v0.1.2
