+++ pkg/util/generics.go
@@ -1,0 +1,34 @@
+/*
+Copyright 2024 Simon Murray
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package util
+
+import (
+	"iter"
+)
+
+// Permute returns an iterator over all possible unique pairs of a slice members.
+func Permute[T any](t []T) iter.Seq2[T, T] {
+	return func(yield func(T, T) bool) {
+		for i := range t {
+			for j := i + 1; j < len(t); j++ {
+				if !yield(t[i], t[j]) {
+					return
+				}
+			}
+		}
+	}
+}
